requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.4.5
django-heroku==0.3.1
django-phone-field==1.8.1
gunicorn==20.1.0
idna==3.2
numpy==1.21.2
pandas==1.3.3
paystack==1.5.0
pep8==1.7.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.2.0
